requirements.txt
requests == 2.26.0
python-dateutil == 2.8.2
termcolor == 1.1.0
colorama == 0.4.4

